+++ app/src/main/java/com/inky/fitnesscalendar/ui/views/settings/ActivityTypeView.kt
@@ -1,9 +1,11 @@
 package com.inky.fitnesscalendar.ui.views.settings
 
+import android.database.sqlite.SQLiteConstraintException
 import androidx.compose.animation.AnimatedContent
 import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.FlowRow
 import androidx.compose.foundation.layout.IntrinsicSize
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.aspectRatio
@@ -21,7 +23,9 @@ import androidx.compose.material3.IconButton
 import androidx.compose.material3.IconButtonDefaults
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
+import androidx.compose.material3.SnackbarDuration
 import androidx.compose.material3.SnackbarHost
+import androidx.compose.material3.SnackbarResult
 import androidx.compose.material3.Switch
 import androidx.compose.material3.Text
 import androidx.compose.material3.TextField
@@ -30,6 +34,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
+import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
@@ -39,10 +44,11 @@ import androidx.compose.ui.platform.LocalResources
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.input.KeyboardCapitalization
 import androidx.compose.ui.unit.dp
-import androidx.hilt.lifecycle.viewmodel.compose.hiltViewModel
 import com.inky.fitnesscalendar.R
 import com.inky.fitnesscalendar.db.entities.ActivityType
+import com.inky.fitnesscalendar.di.AppRepository
 import com.inky.fitnesscalendar.ui.components.ActivityCategorySelector
+import com.inky.fitnesscalendar.ui.components.ActivityTypeFilterChip
 import com.inky.fitnesscalendar.ui.components.ActivityTypeSelector
 import com.inky.fitnesscalendar.ui.components.BaseEditDialog
 import com.inky.fitnesscalendar.ui.components.ColorSelector
@@ -51,12 +57,16 @@ import com.inky.fitnesscalendar.ui.components.OptionGroup
 import com.inky.fitnesscalendar.ui.components.defaultTopAppBarColors
 import com.inky.fitnesscalendar.ui.components.optionGroupDefaultBackground
 import com.inky.fitnesscalendar.ui.util.Icons
+import com.inky.fitnesscalendar.ui.util.localDatabaseValues
 import com.inky.fitnesscalendar.view_model.settings.ActivityTypeEditState
-import com.inky.fitnesscalendar.view_model.settings.ActivityTypeViewModel
+import kotlinx.coroutines.launch
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
-fun ActivityTypeView(viewModel: ActivityTypeViewModel = hiltViewModel(), onBack: () -> Unit) {
+context(app: AppRepository)
+fun ActivityTypeView(onBack: () -> Unit) {
+    val scope = rememberCoroutineScope()
+
     var selectedType by rememberSaveable { mutableStateOf<ActivityType?>(null) }
     val initialEditState = remember(selectedType) {
         selectedType?.let { ActivityTypeEditState(it) } ?: ActivityTypeEditState()
@@ -87,10 +97,10 @@ fun ActivityTypeView(viewModel: ActivityTypeViewModel = hiltViewModel(), onBack:
             )
         },
         snackbarHost = {
-            SnackbarHost(hostState = viewModel.snackbarHostState)
+            SnackbarHost(hostState = app.snackbarHostState)
         },
     ) { paddingValues ->
-        Row(
+        Column(
             modifier = Modifier
                 .padding(paddingValues)
                 .padding(all = 8.dp)
@@ -103,6 +113,10 @@ fun ActivityTypeView(viewModel: ActivityTypeViewModel = hiltViewModel(), onBack:
                     showEditDialog = true
                 },
             )
+
+            ArchivedActivityTypes(
+                onUnarchive = { scope.launch { restore(it) } }
+            )
         }
     }
 
@@ -111,12 +125,16 @@ fun ActivityTypeView(viewModel: ActivityTypeViewModel = hiltViewModel(), onBack:
             initialState = initialEditState,
             onDismiss = { showEditDialog = false },
             onSave = {
-                showEditDialog = false
-                viewModel.save(it)
+                scope.launch {
+                    app.db.saveActivityType(it)
+                    showEditDialog = false
+                }
             },
             onDelete = {
-                showEditDialog = false
-                selectedType?.let { viewModel.delete(it) }
+                scope.launch {
+                    selectedType?.let { delete(it) }
+                    showEditDialog = false
+                }
             }
         )
     }
@@ -151,7 +169,13 @@ fun EditTypeDialog(
         title = title,
         actions = {
             if (!state.isNewType) {
-                EditTypeDialogMenu(onDeleteType = onDelete)
+                EditTypeDialogMenu(
+                    onDeleteType = onDelete,
+                    onArchive = {
+                        type?.let { onSave(it.copy(archived = true)) }
+                    },
+                    showArchiveOption = state == initialState
+                )
             }
         }
     ) {
@@ -272,7 +296,11 @@ fun EditTypeDialog(
 }
 
 @Composable
-fun EditTypeDialogMenu(onDeleteType: () -> Unit) {
+fun EditTypeDialogMenu(
+    onDeleteType: () -> Unit,
+    onArchive: () -> Unit,
+    showArchiveOption: Boolean
+) {
     var showMenu by remember { mutableStateOf(false) }
 
     IconButton(onClick = { showMenu = true }) {
@@ -280,6 +308,13 @@ fun EditTypeDialogMenu(onDeleteType: () -> Unit) {
     }
 
     DropdownMenu(expanded = showMenu, onDismissRequest = { showMenu = false }) {
+        if (showArchiveOption) {
+            DropdownMenuItem(
+                text = { Text(stringResource(R.string.archive_activity_type)) },
+                leadingIcon = { Icons.Archive(stringResource(R.string.archive_activity_type)) },
+                onClick = onArchive
+            )
+        }
         DropdownMenuItem(
             text = { Text(stringResource(R.string.delete_activity_type)) },
             leadingIcon = { Icons.Delete(stringResource(R.string.delete)) },
@@ -313,5 +348,69 @@ fun Toggle(
             onCheckedChange = onValue,
             modifier = Modifier.padding(horizontal = 8.dp)
         )
+    }
+}
+
+@Composable
+fun ArchivedActivityTypes(onUnarchive: (ActivityType) -> Unit) {
+    val activityTypes = localDatabaseValues.current.activityTypes.filter { it.archived }
+
+    AnimatedVisibility(activityTypes.isNotEmpty(), modifier = Modifier.fillMaxWidth()) {
+        Column(
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(vertical = 8.dp)
+                .clip(MaterialTheme.shapes.large)
+                .background(MaterialTheme.colorScheme.surfaceContainer)
+                .padding(all = 8.dp)
+
+        ) {
+            Text(
+                stringResource(R.string.archived_activity_types),
+                style = MaterialTheme.typography.titleLarge,
+                modifier = Modifier.padding(all = 8.dp)
+            )
+            FlowRow {
+                for (type in activityTypes) {
+                    ActivityTypeFilterChip(
+                        activityType = type,
+                        selected = false,
+                        onClick = { onUnarchive(type) },
+                    )
+                }
+            }
+        }
+    }
+}
+
+context(app: AppRepository)
+suspend fun delete(activityType: ActivityType) {
+    try {
+        app.db.deleteActivityType(activityType)
+        val result = app.snackbarHostState.showSnackbar(
+            message = app.context.getString(R.string.deleted_activity_type),
+            actionLabel = app.context.getString(R.string.undo),
+            duration = SnackbarDuration.Short
+        )
+        when (result) {
+            SnackbarResult.ActionPerformed -> app.db.saveActivityType(activityType)
+            SnackbarResult.Dismissed -> {}
+        }
+    } catch (e: SQLiteConstraintException) {
+        app.snackbarHostState.showSnackbar(message = app.context.getString(R.string.cannot_delete_type_because_there_are_still_activities))
+    }
+}
+
+context(app: AppRepository)
+suspend fun restore(activityType: ActivityType) {
+    app.db.saveActivityType(activityType.copy(archived = false))
+    val result = app.snackbarHostState.showSnackbar(
+        message = app.context.getString(R.string.restored_activity_type),
+        actionLabel = app.context.getString(R.string.undo),
+        duration = SnackbarDuration.Short
+    )
+    when (result) {
+        SnackbarResult.ActionPerformed -> app.db.saveActivityType(activityType)
+        SnackbarResult.Dismissed -> {}
     }
 }